requirements.txt
absl-py==1.0.0
altgraph==0.17.2
asttokens==3.0.0
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.1.5
attrs==19.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.0.0
certifi==2019.11.28
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.3
comm==0.2.2
contourpy==1.3.3
cryptography==45.0.6
cycler==0.10.0
Cython==0.29.28
debugpy==1.8.12
decorator==4.4.2
deepface==0.0.95
distro==1.4.0
dnspython==2.8.0
entrypoints==0.3
et_xmlfile==1.0.1
executing==2.2.0
face_recognition_models==0.3.0
filelock==3.19.1
fire==0.7.1
Flask==3.0.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.9.0
gast==0.5.3
gdown==5.2.0
google-auth==2.6.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.74.0
gTTS==2.2.4
gunicorn==23.0.0
h11==0.12.0
h5py==3.14.0
html5lib==1.0.1
httplib2==0.14.0
huggingface-hub==0.34.4
idna==2.8
imageio==2.13.5
importlib-metadata==4.11.3
ipykernel==6.29.5
ipython==9.0.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jdcal==1.4.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jtutils==0.0.8
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.11.3
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.4.9
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
libclang==13.0.0
lz4==4.4.4
macaroonbakery
Markdown==3.3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.3
more-itertools==4.2.0
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
nest-asyncio==1.6.0
networkx==2.6.3
nose==1.3.7
numpy==2.1.3
oauthlib==3.1.0
olefile==0.46
opencv-python==4.10.0.84
openpyxl==3.0.3
opt-einsum==3.3.0
optree==0.17.0
outcome==1.1.0
packaging==21.3
pandas==2.2.3
parso==0.8.4
pbr==7.0.1
pexpect==4.6.0
pillow==11.3.0
platformdirs==4.3.6
playsound==1.2.2
pluggy==0.13.0
prompt_toolkit==3.0.50
protobuf==5.29.5
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pycryptodome==3.23.0
Pygments==2.19.2
pyinstaller-hooks-contrib==2021.4
PyJWT==1.7.1
pymacaroons==0.13.0
pymongo==4.14.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.6
pyRFC3339==1.1
PySocks==1.7.1
pytesseract==0.3.9
pytest==4.6.9
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyWavelets==1.2.0
pywin32==308
pywin32-ctypes==0.2.3
pyxdg==0.26
PyYAML==6.0.2
pyzmq==26.2.1
rave-python==1.4.2
redis==6.4.0
regex==2025.9.1
requests==2.32.5
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
retina-face==0.0.17
rich==14.1.0
rsa==4.8
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
selenium==4.1.0
sentence-transformers==5.1.0
setuptools==80.9.0
six==1.17.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==1.9.5
spotipy==2.25.1
SQLAlchemy==2.0.43
stack-data==0.6.3
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.1
termcolor==3.1.0
testresources==2.0.2
tf-estimator-nightly==2.8.0.dev2021122109
tf_keras==2.19.0
thread6==0.2.0
threadpoolctl==3.6.0
tifffile==2021.11.2
tokenizers==0.22.0
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==1.26.7
urllib3-secure-extra==0.1.0
wadllib==2.0.0
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.0.0
xlrd==1.1.0
xlwt==1.3.0
xmltodict==0.12.0
zipp==1.0.0
